143 command structure, repo structure, and extension structure

+++ Solutions/Enhabit/Enhabit.ViewModels/PriceRangeViewModel.cs
@@ -2,8 +2,8 @@
 {
     public class PriceRangeViewModel
     {
-        public float Low { get; set; }
-        public float High { get; set; }
+        public decimal Low { get; set; }
+        public decimal High { get; set; }
         public int Step { get; set; }
     }
 }
